+++ FitTracker/app/src/main/java/com/example/fittracker/prodotto/ProdottoViewModel.kt
@@ -28,7 +28,7 @@ class ProdottoViewModel : ViewModel() {
 
     fun setPastoOnDB(
         tipologiaPasto: String, foodId: String, image: String, nome: String/*label*/, calorie: Double, proteine: Double,
-        carboidrati: Double, grassi: Double, quantita: Int, context: Context
+        carboidrati: Double, grassi: Double, quantita: Double, context: Context
     ) {
         viewModelScope.launch {
             if (prodottoDB.setPasto(
